refactor: modularize device scanner
