+++ tests/test_optimized_integration.py
@@ -347,7 +347,7 @@ class TestThesslaGreenConfigFlow:
         }
 
         with patch(
-            "custom_components.thessla_green_modbus.device_scanner."
+            "custom_components.thessla_green_modbus.scanner_core."
             "ThesslaGreenDeviceScanner.scan_device",
             return_value=scanner_result,
         ):
@@ -445,9 +445,9 @@ class TestThesslaGreenDeviceScanner:
         return client
 
     @pytest.mark.asyncio
-    async def test_device_scanner_success(self, mock_modbus_client):
+    async def test_scanner_core_success(self, mock_modbus_client):
         """Test successful device scanning."""
-        from custom_components.thessla_green_modbus.device_scanner import (
+        from custom_components.thessla_green_modbus.scanner_core import (
             ThesslaGreenDeviceScanner,
         )
 
@@ -462,9 +462,9 @@ class TestThesslaGreenDeviceScanner:
             assert result["device_info"]["firmware"] == "4.85.2"
 
     @pytest.mark.asyncio
-    async def test_device_scanner_connection_failure(self):
+    async def test_scanner_core_connection_failure(self):
         """Test scanner behavior on connection failure."""
-        from custom_components.thessla_green_modbus.device_scanner import (
+        from custom_components.thessla_green_modbus.scanner_core import (
             ThesslaGreenDeviceScanner,
         )
 
@@ -479,7 +479,7 @@ class TestThesslaGreenDeviceScanner:
 
     def test_register_value_validation(self):
         """Test register value validation logic."""
-        from custom_components.thessla_green_modbus.device_scanner import (
+        from custom_components.thessla_green_modbus.scanner_core import (
             ThesslaGreenDeviceScanner,
         )
 
@@ -518,7 +518,7 @@ class TestThesslaGreenDeviceScanner:
 
     def test_capability_analysis(self):
         """Test capability analysis logic."""
-        from custom_components.thessla_green_modbus.device_scanner import (
+        from custom_components.thessla_green_modbus.scanner_core import (
             ThesslaGreenDeviceScanner,
         )
 
@@ -733,7 +733,7 @@ class TestPerformanceOptimizations:
     @pytest.mark.asyncio
     async def test_scan_optimization_stats(self):
         """Test that device scanner provides optimization statistics."""
-        from custom_components.thessla_green_modbus.device_scanner import (
+        from custom_components.thessla_green_modbus.scanner_core import (
             ThesslaGreenDeviceScanner,
         )
