+++ tests/test_read_cancellation.py
@@ -6,7 +6,7 @@ from unittest.mock import AsyncMock, patch
 
 import pytest
 
-from custom_components.thessla_green_modbus.device_scanner import ThesslaGreenDeviceScanner
+from custom_components.thessla_green_modbus.scanner_core import ThesslaGreenDeviceScanner
 from custom_components.thessla_green_modbus.modbus_exceptions import ModbusIOException
 
 pytestmark = pytest.mark.asyncio
@@ -23,7 +23,7 @@ async def test_read_cancellation_during_sleep(method, caplog):
 
     with (
         patch(
-            "custom_components.thessla_green_modbus.device_scanner._call_modbus",
+            "custom_components.thessla_green_modbus.scanner_core._call_modbus",
             AsyncMock(side_effect=ModbusIOException("boom")),
         ),
         patch("asyncio.sleep", AsyncMock(side_effect=asyncio.CancelledError)),
